Close native MVP review gates
Keep PTY indexing type-safe under x86 -Werror, decode current four-instruction ARM64 import thunks, restore strict unused-function and fallthrough diagnostics, exclude the unavailable locale profile, and package and attest the native Windows ARM64 OpenSSH client.


Copilot-Session: 2cae2cbe-ba84-4289-9716-f4e077b6e1d6

+++ winsup/cygwin/fhandler/pty.cc
@@ -2274,7 +2274,7 @@ fhandler_pty_master::write (const void *ptr, size_t len)
 	 put them together into wpbuf and write all at once. */
       static constexpr size_t wpbuf_len = sizeof ("\033[32768;32868R") - 1;
       static char wpbuf[wpbuf_len];
-      static int ixput = 0;
+      static size_t ixput = 0;
       static int state = 0;
       static DWORD wp_tid = 0;
